Refactor tests to TS
