+++ package.json
@@ -6,10 +6,9 @@
   "types": "src/types.d.ts",
   "scripts": {
     "lint": "eslint --cache \"**/*.{js,ts}\"",
-    "spec": "mocha",
+    "spec": "ts-mocha test/*.ts",
     "test": "npm run lint && npm run spec",
-    "build": "tsc",
-    "watch": "tsc --watch",
+    "build": "tsc -p tsconfig-build.json",
     "prepublishOnly": "npm run build"
   },
   "repository": "ngryman/reading-time",
@@ -24,15 +23,17 @@
   "author": "Nicolas Gryman <ngryman@gmail.com> (http://ngryman.sh)",
   "license": "MIT",
   "devDependencies": {
+    "@types/chai": "^4.2.21",
+    "@types/expect": "^24.3.0",
+    "@types/mocha": "^9.0.0",
     "@types/node": "^16.7.10",
     "@typescript-eslint/eslint-plugin": "^4.30.0",
     "@typescript-eslint/parser": "^4.30.0",
     "chai": "^4.1.0",
-    "curry": "^1.2.0",
     "eslint": "^7.32.0",
     "eslint-config-ngryman": "^1.7.0",
-    "mocha": "^9.0.3",
-    "onchange": "^7.1.0",
+    "mocha": "^8.4.0",
+    "ts-mocha": "^8.0.0",
     "typescript": "^4.4.2"
   }
 }
