+++ tsconfig-build.json
@@ -1,0 +1,12 @@
+{
+  "include": ["src"],
+  "compilerOptions": {
+    "lib": ["esnext"],
+    "outDir": "dist",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "rootDir": "src",
+    "declaration": true,
+    "esModuleInterop": true
+  },
+}
